Feature/bind question

+++ server.py
@@ -32,7 +32,8 @@ def route_add_question():
     if request.method == 'POST':
         new_question = request.form.to_dict()
         new_question_add_to_database = util.initialize_view_number_and_vote_number_and_add_datetime(new_question)
-        question_id = data_manager.add_question(new_question_add_to_database)[0]['id']
+        new_question_add_to_database.update({"site_user_id": session["user_id"]})
+        question_id = data_manager.add_question_to_database_return_its_id(new_question_add_to_database)['id']
         return redirect(url_for('route_question_detail',
                                 question_id=question_id))
     return render_template('/add_question.html')
@@ -42,7 +43,7 @@ def route_add_question():
 @login.login_required
 def route_question_detail(question_id):
     data_manager.increment_view_number(question_id)
-    question_data = data_manager.get_question_data(question_id)[0]
+    question_data = data_manager.get_question_data(question_id)
     answers = data_manager.get_answers_for_question(question_id)
     question_comments = data_manager.get_comments_for_question(question_id)
     answer_comments = data_manager.get_answer_comments()
@@ -55,7 +56,8 @@ def route_question_detail(question_id):
                            answers=answers,
                            question_comments=question_comments,
                            answer_comments=answer_comments,
-                           answer_id_list_for_comments=answer_id_list_for_comments)
+                           answer_id_list_for_comments=answer_id_list_for_comments,
+                           session_user=session["user_name"])
 
 
 @app.route('/question/<question_id>/new-answer', methods=['POST', 'GET'])
@@ -132,22 +134,28 @@ def add_answer_comment(answer_id):
 @app.route('/question/<question_id>/edit', methods=['GET', 'POST'])
 @login.login_required
 def modify_question(question_id):
-    if request.method == 'POST':
-        question_to_update = request.form.to_dict()
-        data_manager.edit_question(question_to_update)
-        return redirect(url_for('route_question_detail',
-                                question_id=question_id))
-    question = data_manager.get_question_data(question_id)[0]
-    return render_template('add_question.html',
-                           question=question,
-                           question_id=question_id)
+    question = data_manager.get_question_data(question_id)
+    if question["user_name"] == session["user_name"]:
+        if request.method == 'POST':
+            question_to_update = request.form.to_dict()
+            data_manager.edit_question(question_to_update)
+            return redirect(url_for('route_question_detail',
+                                    question_id=question_id))
+        return render_template('add_question.html',
+                               question=question,
+                               question_id=question_id)
+    
+    return redirect("https://i.imgflip.com/239qx5.jpg")
 
 
 @app.route('/question/<question_id>/delete', methods=['POST', 'GET'])
 @login.login_required
 def delete_question(question_id):
-    data_manager.delete_question_and_its_answers(question_id)
-    return redirect('/')
+    question = data_manager.get_question_data(question_id)
+    if question["user_name"] == session["user_name"]:
+        data_manager.delete_question_and_its_answers(question_id)
+        return redirect('/')
+    return redirect("https://i.imgflip.com/239qx5.jpg")
 
 
 @app.route('/answer/<answer_id>/edit', methods=['POST', 'GET'])
